[button] Add a brief "active" effect when a button a clicked
This effect is achieved by adding a `-active` class on the button instance when it's clicked, and removing it a few hundreds milliseconds later.
Our 4 default buttons types have default styles (background color is lightened) for this effect, with a transition to make it smoother.

How users can customise it:
 - The visual impact of the `-active` class is pure CSS, and thus can simply be overridden with custom userland CSS
 - The duration for which the `-active` class is added can be customised on each subclass of Button, or even at the Button instance level, by overriding the `ACTIVE_EFFECT_DURATION` property