Extend set of handled reserved characters for screenshots
This extends the set of reserved characters that are guarded against when
auto-generating a screenshot filename. The extension covers the most likely
candidates taken from:

   https://learn.microsoft.com/en-us/windows/win32/fileio/naming-a-file#naming-conventions

This could be spun out into a more general filename cleaning function, which
takes OS and filesystem into account and covers every single reserved
character, but for the purposes of this bit of code this should be more than
enough.

Addresses #993.