Drop support for mounting "named" widgets
Before now it was possible to do something like:

  self.mount(hello=Static("Hello!"))

and the Static widget would be created with the ID of 'hello'. Pretty much
the same as doing:

  self.mount(Static("Hello!", id="hello"))

The former wasn't very well advertised in the documentation, wasn't used in
any example code and wasn't used in internally. Meanwhile it also meant that
it was going to be tricky to add extra parameters to the mount
method (something we want to do for #778).

This commit removes that support.