Add index method to NodeList and use it for sibling-relative inserts
This is close to being overkill, but I also feel that, for the moment
anyway, it's worth having this exist as a reminder that doing an index on
the internal list, to find a widget, could end up getting a bit expensive;
having a single method for performing that should at least highlight that
there's one starting point to improve performance there.

Note that this is only really an issue when doing inserts into the node
list, not when appending.