Tidy the flow through _find_spot
While it was mostly working okay, I felt it needed a wee bit more
protection, and also could stand to have some of the more extreme
type-checking complaints handled. At this point I feel the code is about as
tidy as it can get -- until I fully test it live and find other issues I
guess.

But this does pass the unit tests I've made for it so far.