Change the Remove event so that it takes a list of widgets to remove
This commit sets things up so that a list of widgets can be used, but also
then goes on to ignore the fact that there is a list, only removing the
first one in the list. That's a feature, not a but -- here I'm just testing
that we can make this move and nothing breaks. Next up is to allow removing
all of the widgets while holding off on any layout refresh until all
removals have been performed.