Handle a Remove event that requests multiple widget removals
This finishes off the work to make the loop-free query request to remove
objects work as fast as possible. Removing widgets with that query method
should be more or less as fast as removing a container for the same number
of widgets.

See https://github.com/Textualize/textual/discussions/1110 for some context.