Tidy up the docstring for the checkbox
Google style docstring guidelines say that __init__ args can go either in
the class docstring, or the __init__ docstring; so this class was correct.
But the guidelines also recommend that once one approach has been picked it
should be the same over the whole body of the code. Textual mostly seems to
do it in the __init__ so I'm moving things to here and also adding some
missing args.