FIX css link
This is the part where the [guide](https://textual.textualize.io/guide/widgets/#default-css) explains the `DEFAULT_CSS` feature, leaving a small amount of css in `hello04.css`. However, the App class still links to the css file from the previous version (`hello03.css`). This PR fixes that and updates to the recent one.