Change the default width of a label to auto
See #1315 for the discussion that prompted this, #1316 for the issue where
this particular aspect of it was isolated as something to address.

I'm not 100% convinced that `Static` does the right thing here either, but I
don't want to go changing too make things at once. However, I think a good
argument can be made that `Label` should be `width: auto` by default, this
seems to make perfect sense to me. I think everyone would expect, by
default, that a label's width is the width of the content (plus any
decoration, of course).