Initial porting of textual-keys into the textual CLI
Originally from https://github.com/davep/textual-keys -- please note that
textual-keys was (and still is) released under the GPLv3. As the copyright
holder of that code I am making this version available to, and part of,
Textual, under the same licence as Textual itself.

To be clear: Dave Pearson licences this version of textual-keys as "MIT
License".

See #1342.