Add the keys command to the CHANGELOG, and mention in the docs
It's not really clear to me what a good location in the docs would be for
this -- not every CLI command seems to be mentioned in the devtools for
example, and I'm not sure this quite counts as a devtool vs "a tool for
developers"; colors and easing for example are for developers but aren't
documented as devtools.

So for now I'm going with a mention of it near where bindings are spoken
about.