Add snippets for syntax code blocks.
When defining the formal syntax of CSS rules we will use a syntax similar to that of the MDN docs and we want to be able to link directly from the syntax to the CSS types involved, so we cannot do that from within a markdown code block. Instead, we use explicit HTML that we include in two snippets so that it is easier to maintain.