Expand the docstring for DOMNode.query_one
See #1560. In short: query_one was changed a couple of months back so that
it searched for only a single instance of a hit for the query, rather
than (as before) returning the first item that matched the query.
Unfortunately the docstring didn't get the necessary update.

This should hopefully make things a wee bit clearer in the docs.