Reraise compose TypeError from the error rather than None
In its current form, if someone's `compose` results in an uncaught TypeError
in their own code, the resulting error will hide the source. This small
change ensures that the main problem (compose returned something that can't
be turned into a list) is still reported, but also provides the developer a
trackback they can look at to see what the ultimate source of the issue was.