Seek to improve the explanation of what a percentage does
I've seen a couple or so folk get confused about what a percentage value for
a dimension means, often seeming to think it means that percentage of
the *remaining* space within a container as opposed to a percentage of the
space in the container (note total space, not remaining space).

I'm not sure the wording here is the final form we want (there's more
mention of "dimension" here than your average Star Trek or Doctor Who
episode!), but I think it's narrowing in on it.

Making this change to place-hold the need to improve this and perhaps gather
some input to find the best final wording.