Handle SIGCONT on Un*x systems
The handling of SIGTSTP may be a little more involved in that it seems
Textual is snarfing ^Z before it can result in the relevant signal. This
commit, however, checks that coming back on a SIGCONT can work fine (testing
that suspend works is easily done by sending a SIGTSTP using `kill`).

Right now, with this change, it *almost* works fine. The application comes
back and carries on running fine, with the screen redrawing without any
glitches, and the user can carry on interacting. However, when actually
exiting the application things don't work so well. For example, performing a
^C to exit seems to mostly close down the application, but then it takes
another ^C to finally close the process.

More digging needed...