Add support for Ctrl+Zing an application into suspended mode
For the moment this simply ensures that the usual key combination is caught
and acted upon. When Ctrl+Z is pressed a SIGTSTP is sent to the
application's process.

At the moment I'm not handling this in the drive (that comes next), for the
moment I just wanted to check that this was going to work in terms of
getting the signal to the right place (it does).