Rename the key action tests
There are actions that relate to editing the input too, which I also want to
test, but I'm not minded to lump them all in the same file. So here I'm
renaming to make it clear these tests are just about movement. The editing
ones will come next in their own file.