Add a toggle action to Tree, along with a new binding
See #1433. The idea here is that the user has an option of
expanding/collapsing a non-leaf node without causing a selected event,
or (with auto_expand turned off) cause a selected event without an
expand/collapse event.

As this will need a new binding, I've chosen the space bar as the key to
toggle the expanded state.