Fully exclude the blog files from the offline docs build
The blog files were still getting included and packages, just not linked to
from the resulting "site". This meant that the final tarball was much bigger
than it needed to be, due to animated gifs and movs an the like.

This brings the tarball down from around 27M to 6.4M.