Add a toggle method
Before this I had an action_toggle, which did the work, and also an on_click
which called on it. It seems a bit OTT to add a toggle method too, to then
especially make the other two methods call on this. But my thinking here is
that the developer will be more likely to go looking for a method to do the
toggle, and less likely to go looking for an action, etc.

Also this is more in keeping with the Switch widget's interface.