WiP revamping of how checkboxes and radio buttons look
Having mostly got the background of this working fine, this commit starts to
dial on the actual look and feel we're going to go for. More work to do here
but this is a good save point.

Next to come is to reduce a lot of the (intentional while playing)
complexity of parts of the buttons, and also the styling too.