Ensure the pressed button index is correct in the message
At the time that we're raising RadioSet.Changed, we can't be sure that the
class updates and the like have finished post RadioButton.watch_value
happening; this means that the "public" route to getting the pressed index
could be wrong (or rather, not quite up to date yet).

Meanwhile, here, we know the button involved, so we go looking for its index
directly.