Make Header.show_clock private
Header.show_clock was just a good old plain object attribute, not a
reactive, and also wasn't done as a read-only property. This means someone
could get the impression that it could be modified on the fly and the header
would react.

There might be an argument for turning it into that; but for the moment this
commit makes it private so that nobody gets the wrong impression.