Add some initial rules to a "mega stylesheet"
The start of what will hopefully become a one-stop comprehensive test of all
the fun corners of stylesheet parsing. While not intended to replace all the
other unit tests for CSS, this test should help to quickly and easily be
sure that nothing breaks any valid CSS.

Influenced by #1253 and #1836 this code starts the process of implementing
the requirement outlined in #1838.

Adding more rules -- simple and/or weird -- that are expected and known to
parse without a problem is encouraged.