Create features and CSS monitor lazily.
The point here was to make the attribute 'features' load lazily but the instantiation of the CSS monitor already needed the features to be set, so we make the CSS monitor lazy as well so that we can instantiate the app without having to provide the features right away.

Related issues: #2285, #2288.