Remove the privacy plugin from the offline version of the docs
Yes, this does pretty much undo the "offline" part, but without this the
docs can't build if the builder doesn't have access to the insiders' edition
of mkdocs. Doing this as a first step to investigating if there's any
possible way to achieve what we want here:

- Anyone can build
- They can build offline
- They can build for maximum privacy

If not, this step at least hits the first requirement.

See #2629.