Rework handler processing again to deal with selector fun
I'm like 90% sure about this. This was a "dump it all and work from scratch"
evening experiment, and I think it does what's needed, but I want to read it
back over again in the morning with a post-coffee brain.

Pushing to the repo so I can review then.

Also added lotsa comments so tomorrow me will know what this evening me was
thinking.