Remove unneeded arguments from `DataTable.__init__`
This commit removes the following arguments from `DataTable.__init__`:
  - `show_header`
  - `show_row_labels`
  - `fixed_rows`
  - `fixed_columns`
  - `zebra_stripes`
  - `header_height`
  - `show_cursor`

This allows subclasses of `DataTable` to simply change the default
reactive value on the class level instead of changing the `__init__`
signature or overwriting its behavior at all (which often is not
needed).
Furthermore, this makes the API of the `DataTable` more consistent with
other textual widgets, where reactive attributes such as these ones are
not really exposed via the `__init__` (see for example
`Tree.show_root`).

And finally, these arguments seem not to have been used at all anyways,
since I did not need to update any tests for this to work.